Restructed User Profile
Move User Project and User Group to seperate components
Fix, update for project and groups actions
