+++ frontend/src/topic/UserProfile/components/UserGroup/styles.scss
@@ -1,0 +1,6 @@
+@import '../../../../stylesheets/base';
+@import '../../../../stylesheets/modal';
+
+.groups {
+    grid-area: groups;
+}
