+++ frontend/src/topic/UserProfile/components/UserProject/styles.scss
@@ -1,0 +1,6 @@
+@import '../../../../stylesheets/base';
+@import '../../../../stylesheets/modal';
+
+.projects {
+    grid-area: projects;
+}
