+++ frontend/src/topic/UserProfile/components/index.js
@@ -1,3 +1,5 @@
 export UserEdit from './UserEdit';
-export ProjectAdd from './ProjectAdd';
+export UserProjectAdd from './UserProjectAdd';
+export UserProject from './UserProject';
+export UserGroup from './UserGroup';
 export UserGroupAdd from './UserGroupAdd';
